Changed: - PHPUnit tests folder renamed according to a common convention
